Change pathCompletions to stringCompletions, move more code there (#28245)
